backend/requirements.txt
fastapi
uvicorn
websockets
faster-whisper
smolagents[toolkit]
transformers
torch
numpy
Pillow
python-multipart
soundfile
xxhash
huggingface_hub
accelerate
bitsandbytes
duckduckgo-search
pandas
ddgs
sympy
